refactor(address-form): centralize handlers and buttons, improve reusability
- renamed isUploading state to isLoading.
- renamed EditAddressForm to EditAddress and AddAddressForm to AddAddress.
- created a common AddressForm component for both add/edit features.
- shifted all handlers and form buttons within the AddressForm.
- used useCallback for API call helper functions.
