+++ src/features/user/Settings/EditProfile/AddressManagment/microComponents/AddressFormWrapper.tsx
@@ -1,0 +1,17 @@
+import type { JSX } from 'react';
+import styles from '../AddressManagement.module.scss';
+
+interface Props {
+  children: JSX.Element;
+  label: string;
+}
+const AddressFormLayout = ({ children, label }: Props) => {
+  return (
+    <div className={styles.addressFormContainer}>
+      <h2>{label}</h2>
+      {children}
+    </div>
+  );
+};
+
+export default AddressFormLayout;
